+++ program/orbit_builder.gd
@@ -193,8 +193,8 @@ func make_orbit_from_data(table_name: String, table_row: int, parent: IVBody) ->
 	if _d.ref_plane == "Equatorial":
 		orbit.reference_normal = parent.get_positive_pole()
 	elif _d.ref_plane == "Laplace":
-		var orbit_ra: float = _table_reader.get_real(table_name, "orbit_RA", table_row)
-		var orbit_dec: float = _table_reader.get_real(table_name, "orbit_dec", table_row)
+		var orbit_ra: float = _table_reader.get_float(table_name, "orbit_RA", table_row)
+		var orbit_dec: float = _table_reader.get_float(table_name, "orbit_dec", table_row)
 		orbit.reference_normal = math.convert_spherical2(orbit_ra, orbit_dec)
 		orbit.reference_normal = _ecliptic_rotation * orbit.reference_normal
 		orbit.reference_normal = orbit.reference_normal.normalized()