+++ singletons/global.gd
@@ -30,7 +30,7 @@ extends Node
 const IVOYAGER_VERSION := "0.0.16"
 const IVOYAGER_BUILD := "" # hotfix or debug build
 const IVOYAGER_STATE := "dev" # 'dev', 'alpha', 'beta', 'rc', ''
-const IVOYAGER_YMD := 20230814
+const IVOYAGER_YMD := 20230828
 
 
 # simulator state broadcasts
@@ -108,7 +108,7 @@ var assets := {} # AssetsInitializer loads from dynamic paths specified below
 var settings := {} # IVSettingsManager
 var tables := {} # IVTableImporter; indexed [table_name][field][row_int]
 var enumerations := {} # IVTableImporter; all row names and listed enums (globally unique)
-var precisions := {} # IVTableImporter; indexed as tables but only REAL fields
+var precisions := {} # IVTableImporter; indexed as tables but only FLOAT fields
 var wiki_titles := {} # IVTableImporter; en.wikipedia; TODO: non-en & internal
 var themes := {} # IVThemeManager
 var fonts := {} # IVFontManager
@@ -157,8 +157,8 @@ var vertecies_per_orbit_low_res: int = 100 # for small bodies like asteroids
 var max_camera_distance: float = 5e3 * IVUnits.AU
 var obliquity_of_the_ecliptic := 23.439 * IVUnits.DEG
 var ecliptic_rotation := IVMath.get_x_rotation_matrix(obliquity_of_the_ecliptic)
-var unit_multipliers := IVUnits.MULTIPLIERS
-var unit_functions := IVUnits.FUNCTIONS
+var unit_multipliers := IVUnits.multipliers
+var unit_lambdas := IVUnits.lambdas
 var cache_dir := "user://cache"
 
 # theses could be modified during gameplay