+++ static/units.gd
@@ -21,12 +21,13 @@ class_name IVUnits
 extends Object
 
 # This class defines derived units from base SI units in si_base_units.gd. You
-# should need it only when converting to and from simulator values: i.e., data
-# import, GUI display, and specifying quantities in code.
+# should need it only when converting to and from simulator values: i.e.,
+# specifying quantities in tables or code (to internal) and GUI display (from
+# internal).
 #
 # WE SHOULD NEVER NEED TO CONVERT UNITS IN OUR INTERNAL PROCESSING!
 #
-# See prog_refs/quantity_formatter.gd for generating quantity strings
+# See program/quantity_formatter.gd for display unit strings.
 
 
 # SI base units - all sim units derived from these!
@@ -70,200 +71,211 @@ const TESLA := WEBER / (METER * METER)
 const STANDARD_GM := KM * KM * KM / (SECOND * SECOND) # usually in these units
 const GRAVITATIONAL_CONSTANT := 6.67430e-11 * METER * METER * METER / (KG * SECOND * SECOND)
 
-# Symbols below mostly follow:
+# Unit symbols below mostly follow:
 # https://en.wikipedia.org/wiki/International_System_of_Units
+#
+# We look for unit symbol first in mulitipliers and then in lambdas.
 
-const MULTIPLIERS := {
-	# duplicated symbols have leading underscore(s)
+static var multipliers := {
+	# Duplicated symbols have leading underscore(s).
+	# See IVQuantityFormater for unit display strings.
+	
 	# time
-	"s" : SECOND,
-	"min" : MINUTE,
-	"h" : HOUR,
-	"d" : DAY,
-	"a" : YEAR, # official Julian year symbol
-	"y" : YEAR,
-	"yr" : YEAR,
-	"Cy" : CENTURY,
+	&"s" : SECOND,
+	&"min" : MINUTE,
+	&"h" : HOUR,
+	&"d" : DAY,
+	&"a" : YEAR, # official Julian year symbol
+	&"y" : YEAR,
+	&"yr" : YEAR,
+	&"Cy" : CENTURY,
 	# length
-	"mm" : MM,
-	"cm" : CM,
-	"m" : METER,
-	"km" : KM,
-	"au" : AU,
-	"AU" : AU,
-	"ly" : LIGHT_YEAR,
-	"pc" : PARSEC,
-	"Mpc" : 1e6 * PARSEC,
+	&"mm" : MM,
+	&"cm" : CM,
+	&"m" : METER,
+	&"km" : KM,
+	&"au" : AU,
+	&"AU" : AU,
+	&"ly" : LIGHT_YEAR,
+	&"pc" : PARSEC,
+	&"Mpc" : 1e6 * PARSEC,
 	# mass
-	"g" : GRAM,
-	"kg" : KG,
-	"t" : TONNE,
+	&"g" : GRAM,
+	&"kg" : KG,
+	&"t" : TONNE,
 	# angle
-	"rad" : 1.0,
-	"deg" : DEG,
+	&"rad" : 1.0,
+	&"deg" : DEG,
 	# temperature
-	"K" : KELVIN,
+	&"K" : KELVIN,
 	# frequency
-	"Hz" : 1.0 / SECOND,
-	"d^-1" : 1.0 / DAY,
-	"a^-1" : 1.0 / YEAR,
-	"y^-1" : 1.0 / YEAR,
-	"yr^-1" : 1.0 / YEAR,
+	&"Hz" : 1.0 / SECOND,
+	&"d^-1" : 1.0 / DAY,
+	&"a^-1" : 1.0 / YEAR,
+	&"y^-1" : 1.0 / YEAR,
+	&"yr^-1" : 1.0 / YEAR,
 	# area
-	"m^2" : METER * METER,
-	"km^2" : KM * KM,
-	"ha" : HECTARE,
+	&"m^2" : METER * METER,
+	&"km^2" : KM * KM,
+	&"ha" : HECTARE,
 	# volume
-	"l" : LITER,
-	"L" : LITER,
-	"m^3" : METER * METER * METER,
-	"km^3" : KM * KM * KM,
+	&"l" : LITER,
+	&"L" : LITER,
+	&"m^3" : METER * METER * METER,
+	&"km^3" : KM * KM * KM,
 	# velocity
-	"m/s" : METER / SECOND,
-	"km/s" : KM / SECOND,
-	"km/h" : KM / HOUR,
-	"au/a" : AU / YEAR,
-	"au/Cy" : AU / CENTURY,
-	"AU/Cy" : AU / CENTURY,
-	"c" : SPEED_OF_LIGHT,
+	&"m/s" : METER / SECOND,
+	&"km/s" : KM / SECOND,
+	&"km/h" : KM / HOUR,
+	&"au/a" : AU / YEAR,
+	&"au/Cy" : AU / CENTURY,
+	&"AU/Cy" : AU / CENTURY,
+	&"c" : SPEED_OF_LIGHT,
 	# acceleration/gravity
-	"m/s^2" : METER / (SECOND * SECOND),
-	"_g" : STANDARD_GRAVITY,
+	&"m/s^2" : METER / (SECOND * SECOND),
+	&"_g" : STANDARD_GRAVITY,
 	# angular velocity
-	"rad/s" : 1.0 / SECOND, 
-	"deg/d" : DEG / DAY,
-	"deg/a" : DEG / YEAR,
-	"deg/Cy" : DEG / CENTURY,
+	&"rad/s" : 1.0 / SECOND, 
+	&"deg/d" : DEG / DAY,
+	&"deg/a" : DEG / YEAR,
+	&"deg/Cy" : DEG / CENTURY,
 	# particle density
-	"m^-3" : 1.0 / (METER * METER * METER),
+	&"m^-3" : 1.0 / (METER * METER * METER),
 	# density
-	"kg/km^3" : KG / (KM * KM * KM),
-	"g/cm^3" : GRAM / (CM * CM * CM),
+	&"kg/km^3" : KG / (KM * KM * KM),
+	&"g/cm^3" : GRAM / (CM * CM * CM),
 	# mass rate
-	"kg/s" : KG / SECOND,
-	"g/d" : GRAM / DAY,
-	"kg/d" : KG / DAY,
-	"t/d" : TONNE / DAY,
+	&"kg/s" : KG / SECOND,
+	&"g/d" : GRAM / DAY,
+	&"kg/d" : KG / DAY,
+	&"t/d" : TONNE / DAY,
 	# force
-	"N" : NEWTON,
+	&"N" : NEWTON,
 	# pressure
-	"Pa" : PASCAL,
-	"bar" : BAR,
-	"atm" : ATM,
+	&"Pa" : PASCAL,
+	&"bar" : BAR,
+	&"atm" : ATM,
 	# energy
-	"J" : JOULE,
-	"kJ" : 1e3 * JOULE,
-	"MJ" : 1e6 * JOULE,
-	"GJ" : 1e9 * JOULE,
-	"TJ" : 1e12 * JOULE,
-	"Wh" : WATT * HOUR,
-	"kWh" : 1e3 * WATT * HOUR,
-	"MWh" : 1e6 * WATT * HOUR,
-	"GWh" : 1e9 * WATT * HOUR,
-	"TWh" : 1e12 * WATT * HOUR,
-	"eV" : ELECTRONVOLT,
+	&"J" : JOULE,
+	&"kJ" : 1e3 * JOULE,
+	&"MJ" : 1e6 * JOULE,
+	&"GJ" : 1e9 * JOULE,
+	&"TJ" : 1e12 * JOULE,
+	&"Wh" : WATT * HOUR,
+	&"kWh" : 1e3 * WATT * HOUR,
+	&"MWh" : 1e6 * WATT * HOUR,
+	&"GWh" : 1e9 * WATT * HOUR,
+	&"TWh" : 1e12 * WATT * HOUR,
+	&"eV" : ELECTRONVOLT,
 	# power
-	"W" : WATT,
-	"kW" : 1e3 * WATT,
-	"MW" : 1e6 * WATT,
-	"GW" : 1e9 * WATT,
-	"TW" : 1e12 * WATT,
-	"GJ/d" : 1e9 * JOULE / DAY,
+	&"W" : WATT,
+	&"kW" : 1e3 * WATT,
+	&"MW" : 1e6 * WATT,
+	&"GW" : 1e9 * WATT,
+	&"TW" : 1e12 * WATT,
+	&"GJ/d" : 1e9 * JOULE / DAY,
 	# luminous intensity / luminous flux
-	"cd" : CANDELA,
-	"cd sr" : CANDELA, # sr is dimentionless
-	"lm" : CANDELA, # lumen
+	&"cd" : CANDELA,
+	&"cd sr" : CANDELA, # sr is dimentionless
+	&"lm" : CANDELA, # lumen
 	# luminance
-	"cd/m^2" : CANDELA / (METER * METER),
+	&"cd/m^2" : CANDELA / (METER * METER),
 	# electric potential
-	"V" : VOLT,
+	&"V" : VOLT,
 	# electric charge
-	"C" :  COULOMB,
+	&"C" :  COULOMB,
 	# magnetic flux
-	"Wb" : WEBER,
+	&"Wb" : WEBER,
 	# magnetic flux density
-	"T" : TESLA,
+	&"T" : TESLA,
 	# GM
-	"km^3/s^2" : STANDARD_GM,
-	"m^3/s^2" : METER * METER * METER / (SECOND * SECOND),
+	&"km^3/s^2" : STANDARD_GM,
+	&"m^3/s^2" : METER * METER * METER / (SECOND * SECOND),
 	# gravitational constant
-	"m^3/(kg s^2)" : METER * METER * METER / (KG * SECOND * SECOND),
-	"km^3/(kg s^2)" : KM * KM * KM / (KG * SECOND * SECOND),
+	&"m^3/(kg s^2)" : METER * METER * METER / (KG * SECOND * SECOND),
+	&"km^3/(kg s^2)" : KM * KM * KM / (KG * SECOND * SECOND),
 	# information (base 10; KiB, MiB, etc. would take some coding...)
-	"bit" : 1.0,
-	"b" : 1.0,
-	"kb" : 1e3,
-	"Mb" : 1e6,
-	"Gb" : 1e9,
-	"Tb" : 1e12,
-	"Byte" : 8.0,
-	"B" : 8.0,
-	"kB" : 8e3,
-	"MB" : 8e6,
-	"GB" : 8e9,
-	"TB" : 8e12,
+	&"bit" : 1.0,
+	&"b" : 1.0,
+	&"kb" : 1e3,
+	&"Mb" : 1e6,
+	&"Gb" : 1e9,
+	&"Tb" : 1e12,
+	&"Byte" : 8.0,
+	&"B" : 8.0,
+	&"kB" : 8e3,
+	&"MB" : 8e6,
+	&"GB" : 8e9,
+	&"TB" : 8e12,
 	# misc
-	"deg/Cy^2" : DEG / (CENTURY * CENTURY),
+	&"deg/Cy^2" : DEG / (CENTURY * CENTURY),
 }
 
-const FUNCTIONS := {
-	# TODO34: use lambdas
-	"degC" : "convert_centigrade",
-	"degF" : "convert_fahrenheit",
+static var lambdas := {
+	&"degC" : func convert_centigrade(x: float, to_internal := true) -> float:
+		return x + 273.15 if to_internal else x - 273.15,
+	&"degF" : func convert_fahrenheit(x: float, to_internal := true) -> float:
+		return  (x + 459.67) / 1.8 if to_internal else x * 1.8 - 459.67,
 }
 
 
-static func convert_quantity(x: float, unit: String, to_internal := true,
-		preprocess := false, multipliers := MULTIPLIERS, functions := FUNCTIONS) -> float:
+
+static func convert_quantity(x: float, unit: StringName, to_internal := true,
+		multipliers_ := multipliers, lambdas_ := lambdas, handle_unit_prefix := false) -> float:
 	# Converts x in specified units to internal representation (to_internal =
 	# true) or from internal to specified units (to_internal = false).
-	# unit can be any key in MULTIPLIERS or FUNCTIONS (or supplied replacement
-	# dictionaries); preprocess = true handles prefixes "10^x " or "1/".
-	# Valid examples: "1/Cy", "10^24 kg", "1/(10^3 yr)".
-	if preprocess: # mainly for table import
+	#
+	# If handle_unit_prefix == true, we handle simple unit prefixes '10^x ' and
+	# '1/'. Valid examples: "1/Cy", "10^24 kg", "1/(10^3 yr)".
+	#
+	# After prefix handling (if used), 'unit' must be a dictionary key in either
+	# 'multipliers_' or 'lambdas_'.
+	
+	if handle_unit_prefix:
 		if unit.begins_with("1/"):
-			unit = unit.trim_prefix("1/")
-			if unit.begins_with("(") and unit.ends_with(")"):
-				unit = unit.lstrip("(").rstrip(")")
+			var unit_str := unit.trim_prefix("1/")
+			if unit_str.begins_with("(") and unit_str.ends_with(")"):
+				unit_str = unit_str.trim_prefix("(").trim_suffix(")")
+			unit = StringName(unit_str)
 			to_internal = !to_internal
 		if unit.begins_with("10^"):
-			var space_pos := unit.find(" ")
-			assert(space_pos > 3, "A space must follow '10^xx'")
-			var exponent_str := unit.substr(3, space_pos - 3)
-			var pre_multiplier := pow(10.0, int(exponent_str))
-			unit = unit.substr(space_pos + 1, 999)
+			var unit_str := unit.trim_prefix("10^")
+			var space_pos := unit_str.find(" ")
+			assert(space_pos > 0, "A space must follow '10^xx'")
+			var exponent_str := unit_str.substr(0, space_pos)
+			assert(exponent_str.is_valid_int())
+			var pre_multiplier := 10.0 ** exponent_str.to_int()
+			unit_str = unit_str.substr(space_pos + 1, 999)
+			unit = StringName(unit_str)
 			x *= pre_multiplier
-	var multiplier: float = multipliers.get(unit, 0.0)
+	
+	var multiplier: float = multipliers_.get(unit, 0.0)
 	if multiplier:
 		return x * multiplier if to_internal else x / multiplier
-	if functions.has(unit):
-		# TODO 4.0: fix this hack when we have 1st class functions!
-		var units = load("res://ivoyager/static/units.gd")
-		return units.call(functions[unit], x, to_internal)
-	assert(false, "Unknown unit symbol: " + unit)
-	return x
+	assert(lambdas_.has(unit), "Unknown unit symbol '%s'" % unit)
+	var lambda: Callable = lambdas_[unit]
+	return lambda.call(x, to_internal)
 
 
-static func convert_centigrade(x: float, to_internal := true) -> float:
-	return x + 273.15 if to_internal else x - 273.15
-
-
-static func convert_fahrenheit(x: float, to_internal := true) -> float:
-	return  (x + 459.67) * (5.0 / 9.0) if to_internal else x * (9.0 / 5.0) - 459.67
-
-
-static func is_valid_unit(unit: String, preprocess := false,
-		multipliers := MULTIPLIERS, functions := FUNCTIONS) -> bool:
-	# unit can be any key in MULTIPLIERS or FUNCTIONS (or supplied replacement
-	# dictionaries); preprocess = true handles prefixes "10^x " or "1/".
-	# Valid examples: "1/Cy", "10^24 kg", "1/(10^3 yr)".
-	if preprocess: # mainly for table import
+static func is_valid_unit(unit: StringName, multipliers_ := multipliers, lambdas_ := lambdas,
+		handle_unit_prefix := false) -> bool:
+	# Tests whether 'unit' string is valid for convert_quantity().
+	if handle_unit_prefix:
 		if unit.begins_with("1/"):
-			unit = unit.trim_prefix("1/")
-			if unit.begins_with("(") and unit.ends_with(")"):
-				unit = unit.lstrip("(").rstrip(")")
+			var unit_str := unit.trim_prefix("1/")
+			if unit_str.begins_with("(") and unit_str.ends_with(")"):
+				unit_str = unit_str.trim_prefix("(").trim_suffix(")")
+			unit = StringName(unit_str)
 		if unit.begins_with("10^"):
-			var space_pos := unit.find(" ")
-			assert(space_pos > 3, "A space must follow '10^xx'")
-			unit = unit.substr(space_pos + 1, 999)
-	return multipliers.has(unit) or functions.has(unit)
+			var unit_str := unit.trim_prefix("10^")
+			var space_pos := unit_str.find(" ")
+			if space_pos <= 0:
+				return false
+			var exponent_str := unit_str.substr(0, space_pos)
+			if !exponent_str.is_valid_int():
+				return false
+			unit_str = unit_str.substr(space_pos + 1, 999)
+			unit = StringName(unit_str)
+	
+	return multipliers_.has(unit) or lambdas_.has(unit)
+